Restructure flask modules
We want to have functions separated from app init so that they can be imported for testing
