+++ api/untitledapp/politics.py
@@ -11,9 +11,10 @@ import uuid
 import flask
 import flask_praetorian
 
+import untitledapp.misc as uam
 import untitledapp.getters as uag
 import untitledapp.shared as uas
-from untitledapp import app, alive_required, start_required, REQUESTS_SESSION, SOCK_HANDLERS, acquire_lock, acquire_locks, release_lock, release_locks_by_id, release_locks_by_name
+from untitledapp import app, alive_required, start_required, REQUESTS_SESSION, SOCK_HANDLERS
 
 @app.route('/api/galaxypolitics/leader', methods=['POST'])
 @flask_praetorian.auth_required
@@ -25,7 +26,7 @@ def galaxy_leader():
     
     galaxy_votes, galaxy_id = uag._get_galaxy_politics(kd_id)
     
-    if not acquire_lock(f'/galaxy/{galaxy_id}/politics'):
+    if not uam.acquire_lock(f'/galaxy/{galaxy_id}/politics'):
         return (flask.jsonify({"message": "Server is busy"}), 400)
     
     try:
@@ -53,7 +54,7 @@ def galaxy_leader():
             data=json.dumps(patch_payload)
         )
     finally:
-        release_lock(f'/galaxy/{galaxy_id}/politics')
+        uam.release_lock(f'/galaxy/{galaxy_id}/politics')
 
     return (flask.jsonify(galaxy_votes), 200)
 
@@ -67,7 +68,7 @@ def galaxy_policies():
     
     galaxy_votes, galaxy_id = uag._get_galaxy_politics(kd_id)
     
-    if not acquire_lock(f'/galaxy/{galaxy_id}/politics'):
+    if not uam.acquire_lock(f'/galaxy/{galaxy_id}/politics'):
         return (flask.jsonify({"message": "Server is busy"}), 400)
 
     try:
@@ -99,7 +100,7 @@ def galaxy_policies():
             data=json.dumps(patch_payload)
         )
     finally:
-        release_lock(f'/galaxy/{galaxy_id}/politics')
+        uam.release_lock(f'/galaxy/{galaxy_id}/politics')
 
     return (flask.jsonify(galaxy_votes), 200)
 
@@ -137,7 +138,7 @@ def create_empire():
     if not valid_name:
         return flask.jsonify({"message": message}), 400
     
-    if not acquire_lock(f'/empires'):
+    if not uam.acquire_lock(f'/empires'):
         return (flask.jsonify({"message": "Server is busy"}), 400)
     
     try:
@@ -152,7 +153,7 @@ def create_empire():
             data=json.dumps(empire_payload)
         )
     finally:
-        release_lock('/empires')
+        uam.release_lock('/empires')
 
     return flask.jsonify({"message": "Empire created", "status": "success"}), 201
 
@@ -177,7 +178,7 @@ def request_join_empire(target_empire):
     empires_inverted, _, _, _ = uag._get_empires_inverted()
 
     request_id = str(uuid.uuid4())
-    if not acquire_locks([f'/empire/{target_empire}/politics', f'/galaxy/{kd_galaxy}/politics'], request_id=request_id):
+    if not uam.acquire_locks([f'/empire/{target_empire}/politics', f'/galaxy/{kd_galaxy}/politics'], request_id=request_id):
         return (flask.jsonify({"message": "Server is busy"}), 400)
 
     try:
@@ -213,7 +214,7 @@ def request_join_empire(target_empire):
             data=json.dumps(galaxy_payload)
         )
     finally:
-        release_locks_by_id(request_id)
+        uam.release_locks_by_id(request_id)
     return flask.jsonify({"message": "Join request sent", "status": "success"}), 200
 
 @app.route('/api/empire/<target_empire>/canceljoin', methods=['POST'])
@@ -230,7 +231,7 @@ def cancel_join_empire(target_empire):
         return flask.jsonify({"message": "You must be galaxy leader to manage Empire requests"}), 400    
     
     request_id = str(uuid.uuid4())
-    if not acquire_locks([f'/empire/{target_empire}/politics', f'/galaxy/{kd_galaxy}/politics'], request_id=request_id):
+    if not uam.acquire_locks([f'/empire/{target_empire}/politics', f'/galaxy/{kd_galaxy}/politics'], request_id=request_id):
         return (flask.jsonify({"message": "Server is busy"}), 400)
 
     try:
@@ -261,7 +262,7 @@ def cancel_join_empire(target_empire):
             data=json.dumps(galaxy_payload)
         )
     finally:
-        release_locks_by_id(request_id)
+        uam.release_locks_by_id(request_id)
     return flask.jsonify({"message": "Join request cancelled", "status": "success"}), 200
 
 def _validate_empire_invite(galaxy_politics, kd_id, empires_inverted):    
@@ -285,7 +286,7 @@ def accept_empire_invite(target_empire):
     empires_inverted, empires_info, _, _ = uag._get_empires_inverted()
 
     request_id = str(uuid.uuid4())
-    if not acquire_locks(['/empires', f'/empire/{target_empire}/politics', f'/galaxy/{kd_galaxy}/politics'], request_id=request_id):
+    if not uam.acquire_locks(['/empires', f'/empire/{target_empire}/politics', f'/galaxy/{kd_galaxy}/politics'], request_id=request_id):
         return (flask.jsonify({"message": "Server is busy"}), 400)
 
     try:
@@ -330,7 +331,7 @@ def accept_empire_invite(target_empire):
             data=json.dumps(galaxy_payload)
         )
     finally:
-        release_locks_by_id(request_id)
+        uam.release_locks_by_id(request_id)
     return flask.jsonify({"message": "Joined Empire", "status": "success"}), 200
 
 def _validate_invite_galaxy(empire_politics, kd_id, galaxy_empires, galaxy_id, kd_galaxy_politics, kd_galaxy_id):    
@@ -359,7 +360,7 @@ def invite_galaxy(target_galaxy):
     kd_empire = empires_inverted.get(kd_id)
 
     request_id = str(uuid.uuid4())
-    if not acquire_locks([f'/empire/{kd_empire}/politics', f'/galaxy/{target_galaxy}/politics'], request_id=request_id):
+    if not uam.acquire_locks([f'/empire/{kd_empire}/politics', f'/galaxy/{target_galaxy}/politics'], request_id=request_id):
         return (flask.jsonify({"message": "Server is busy"}), 400)
     
     try:
@@ -395,7 +396,7 @@ def invite_galaxy(target_galaxy):
             data=json.dumps(galaxy_payload)
         )
     finally:
-        release_locks_by_id(request_id)
+        uam.release_locks_by_id(request_id)
     return flask.jsonify({"message": "Invitation sent", "status": "success"}), 200
 
 @app.route('/api/galaxy/<target_galaxy>/cancelinvite', methods=['POST'])
@@ -412,7 +413,7 @@ def cancel_invite_galaxy(target_galaxy):
     kd_empire = empires_inverted.get(kd_id)
 
     request_id = str(uuid.uuid4())
-    if not acquire_locks([f'/empire/{kd_empire}/politics', f'/galaxy/{target_galaxy}/politics'], request_id=request_id):
+    if not uam.acquire_locks([f'/empire/{kd_empire}/politics', f'/galaxy/{target_galaxy}/politics'], request_id=request_id):
         return (flask.jsonify({"message": "Server is busy"}), 400)
     
     try:
@@ -450,7 +451,7 @@ def cancel_invite_galaxy(target_galaxy):
             data=json.dumps(galaxy_payload)
         )
     finally:
-        release_locks_by_id(request_id)
+        uam.release_locks_by_id(request_id)
     return flask.jsonify({"message": "Invitation revoked", "status": "success"}), 200
 
 def _validate_accept_galaxy_request(empire_politics, kd_id, galaxy_empires, galaxy_id, kd_galaxy_politics, kd_galaxy_id):    
@@ -479,7 +480,7 @@ def accept_galaxy_request(target_galaxy):
     kd_empire = empires_inverted.get(kd_id)
 
     request_id = str(uuid.uuid4())
-    if not acquire_locks([f'/empire/{kd_empire}/politics', f'/galaxy/{target_galaxy}/politics'], request_id=request_id):
+    if not uam.acquire_locks([f'/empire/{kd_empire}/politics', f'/galaxy/{target_galaxy}/politics'], request_id=request_id):
         return (flask.jsonify({"message": "Server is busy"}), 400)
 
     try:
@@ -525,7 +526,7 @@ def accept_galaxy_request(target_galaxy):
             data=json.dumps(galaxy_payload)
         )
     finally:
-        release_locks_by_id(request_id)
+        uam.release_locks_by_id(request_id)
     return flask.jsonify({"message": "Galaxy added to Empire", "status": "success"}), 200
 
 def _validate_leave_empire(empire_politics, kd_id, kd_galaxy_politics):
@@ -548,7 +549,7 @@ def leave_empire():
     kd_empire = empires_inverted.get(kd_id)
 
     request_id = str(uuid.uuid4())
-    if not acquire_locks(['/empires', f'/empire/{kd_empire}/politics'], request_id=request_id):
+    if not uam.acquire_locks(['/empires', f'/empire/{kd_empire}/politics'], request_id=request_id):
         return (flask.jsonify({"message": "Server is busy"}), 400)
     
     try:
@@ -584,7 +585,7 @@ def leave_empire():
                 data=json.dumps(kd_empire_payload)
             )
     finally:
-        release_locks_by_id(request_id)
+        uam.release_locks_by_id(request_id)
     return flask.jsonify({"message": "Left Empire", "status": "success"}), 200
 
 def _validate_denounce(empire_politics, kd_id, kd_galaxy_politics, kd_galaxy_id, empires_info, target_empire, kd_empire):    
@@ -615,7 +616,7 @@ def denounce(target_empire):
     kd_empire = empires_inverted.get(kd_id)
 
     request_id = str(uuid.uuid4())
-    if not acquire_locks(['/universenews', '/empires', f'/empire/{kd_empire}/politics'], request_id=request_id):
+    if not uam.acquire_locks(['/universenews', '/empires', f'/empire/{kd_empire}/politics'], request_id=request_id):
         return (flask.jsonify({"message": "Server is busy"}), 400)
 
     try:
@@ -660,7 +661,7 @@ def denounce(target_empire):
             data=json.dumps(empires_payload)
         )
     finally:
-        release_locks_by_id(request_id)
+        uam.release_locks_by_id(request_id)
     return flask.jsonify({"message": "Denounced!", "status": "success"}), 200
 
 def _validate_declare_war(empire_politics, kd_id, kd_galaxy_politics, kd_galaxy_id, empires_info, target_empire, kd_empire):    
@@ -694,7 +695,7 @@ def declare_war(target_empire):
     kd_empire = empires_inverted.get(kd_id)
 
     request_id = str(uuid.uuid4())
-    if not acquire_locks(['/universenews', '/empires'], request_id=request_id):
+    if not uam.acquire_locks(['/universenews', '/empires'], request_id=request_id):
         return (flask.jsonify({"message": "Server is busy"}), 400)
 
     try:
@@ -735,7 +736,7 @@ def declare_war(target_empire):
             data=json.dumps(empires_payload)
         )
     finally:
-        release_locks_by_id(request_id)
+        uam.release_locks_by_id(request_id)
     return flask.jsonify({"message": "Declared war!", "status": "success"}), 200
 
 def _validate_request_surrender(empire_politics, kd_id, kd_galaxy_politics, kd_galaxy_id, empires_info, target_empire, kd_empire, surrender_type, surrender_value):    
@@ -772,7 +773,7 @@ def request_surrender(target_empire):
     kd_empire = empires_inverted.get(kd_id)
 
     request_id = str(uuid.uuid4())
-    if not acquire_locks([f'/empire/{target_empire}/politics', f'/empire/{kd_empire}/politics'], request_id=request_id):
+    if not uam.acquire_locks([f'/empire/{target_empire}/politics', f'/empire/{kd_empire}/politics'], request_id=request_id):
         return (flask.jsonify({"message": "Server is busy"}), 400)
 
     try:
@@ -836,7 +837,7 @@ def request_surrender(target_empire):
             data=json.dumps(empire_payload)
         )
     finally:
-        release_locks_by_id(request_id)
+        uam.release_locks_by_id(request_id)
 
     return flask.jsonify({"message": "Surrender request sent", "status": "success"}), 200
 
@@ -856,7 +857,7 @@ def cancel_request_surrender(target_empire):
     kd_empire = empires_inverted.get(kd_id)
 
     request_id = str(uuid.uuid4())
-    if not acquire_locks([f'/empire/{target_empire}/politics', f'/empire/{kd_empire}/politics'], request_id=request_id):
+    if not uam.acquire_locks([f'/empire/{target_empire}/politics', f'/empire/{kd_empire}/politics'], request_id=request_id):
         return (flask.jsonify({"message": "Server is busy"}), 400)
     
     try:
@@ -916,7 +917,7 @@ def cancel_request_surrender(target_empire):
             data=json.dumps(empire_payload)
         )
     finally:
-        release_locks_by_id(request_id)
+        uam.release_locks_by_id(request_id)
 
     return flask.jsonify({"message": "Surrender request cancelled", "status": "success"}), 200
 
@@ -954,7 +955,7 @@ def offer_surrender(target_empire):
     kd_empire = empires_inverted.get(kd_id)
 
     request_id = str(uuid.uuid4())
-    if not acquire_locks([f'/empire/{target_empire}/politics', f'/empire/{kd_empire}/politics'], request_id=request_id):
+    if not uam.acquire_locks([f'/empire/{target_empire}/politics', f'/empire/{kd_empire}/politics'], request_id=request_id):
         return (flask.jsonify({"message": "Server is busy"}), 400)
     
     try:
@@ -1018,7 +1019,7 @@ def offer_surrender(target_empire):
             data=json.dumps(empire_payload)
         )
     finally:
-        release_locks_by_id(request_id)
+        uam.release_locks_by_id(request_id)
 
     return flask.jsonify({"message": "Surrender offer sent", "status": "success"}), 200
 
@@ -1038,7 +1039,7 @@ def cancel_offer_surrender(target_empire):
     kd_empire = empires_inverted.get(kd_id)
 
     request_id = str(uuid.uuid4())
-    if not acquire_locks([f'/empire/{target_empire}/politics', f'/empire/{kd_empire}/politics'], request_id=request_id):
+    if not uam.acquire_locks([f'/empire/{target_empire}/politics', f'/empire/{kd_empire}/politics'], request_id=request_id):
         return (flask.jsonify({"message": "Server is busy"}), 400)
 
     try:
@@ -1098,7 +1099,7 @@ def cancel_offer_surrender(target_empire):
             data=json.dumps(empire_payload)
         )
     finally:
-        release_locks_by_id(request_id)
+        uam.release_locks_by_id(request_id)
 
     return flask.jsonify({"message": "Surrender offer cancelled", "status": "success"}), 200
 
@@ -1247,7 +1248,7 @@ def accept_offer_surrender(target_empire):
     base_locks.extend([
         f"/kingdom/{kd_id}" for kd_id in winning_kds
     ])
-    if not acquire_locks(base_locks, request_id=request_id):
+    if not uam.acquire_locks(base_locks, request_id=request_id):
         return (flask.jsonify({"message": "Server is busy"}), 400)
     
     try:
@@ -1285,7 +1286,7 @@ def accept_offer_surrender(target_empire):
             surrender_value,
         )
     finally:
-        release_locks_by_id(request_id)
+        uam.release_locks_by_id(request_id)
     return flask.jsonify({"message": "Surrender offer has been accepted", "status": "success"}), 200
 
 @app.route('/api/empire/<target_empire>/acceptsurrenderrequest', methods=['POST'])
@@ -1320,7 +1321,7 @@ def accept_request_surrender(target_empire):
     base_locks.extend([
         f"/kingdom/{kd_id}" for kd_id in winning_kds
     ])
-    if not acquire_locks(base_locks, request_id=request_id):
+    if not uam.acquire_locks(base_locks, request_id=request_id):
         return (flask.jsonify({"message": "Server is busy"}), 400)
 
     try:
@@ -1358,7 +1359,7 @@ def accept_request_surrender(target_empire):
             surrender_value,
         )
     finally:
-        release_locks_by_id(request_id)
+        uam.release_locks_by_id(request_id)
     return flask.jsonify({"message": "Surrender request has been accepted", "status": "success"}), 200
     
 
@@ -1384,7 +1385,7 @@ def buy_votes():
     req = flask.request.get_json(force=True)
     kd_id = flask_praetorian.current_user().kd_id
     
-    if not acquire_lock(f'/kingdom/{kd_id}'):
+    if not uam.acquire_lock(f'/kingdom/{kd_id}'):
         return (flask.jsonify({"message": "Server is busy"}), 400)
     try:
         kd_info = uag._get_kd_info(kd_id)
@@ -1406,7 +1407,7 @@ def buy_votes():
             data=json.dumps(kd_patch_payload)
         )
     finally:
-        release_lock(f'/kingdom/{kd_id}')
+        uam.release_lock(f'/kingdom/{kd_id}')
     return (flask.jsonify({"message": "Bought votes", "status": "success"}), 200)
 
 def _validate_cast_votes(
@@ -1430,7 +1431,7 @@ def universe_policies():
     kd_id = flask_praetorian.current_user().kd_id
     
     request_id = str(uuid.uuid4())
-    if not acquire_locks([f'/kingdom/{kd_id}', f'/universepolitics'], request_id=request_id):
+    if not uam.acquire_locks([f'/kingdom/{kd_id}', f'/universepolitics'], request_id=request_id):
         return (flask.jsonify({"message": "Server is busy"}), 400)
     
     try:
@@ -1467,6 +1468,6 @@ def universe_policies():
             data=json.dumps(universe_politics)
         )
     finally:
-        release_locks_by_id(request_id)
+        uam.release_locks_by_id(request_id)
 
     return (flask.jsonify({"message": "Cast votes", "status": "success"}), 200)
